Statistics repository

+++ src/Lykke.Job.BlockchainRiskControl.Domain/Repositories/IStatisticsRepository.cs
@@ -5,8 +5,8 @@ namespace Lykke.Job.BlockchainRiskControl.Domain.Repositories
 {
     public interface IStatisticsRepository
     {
-        Task<int> GetOperationsCountForTheLastPeriodAsync(TimeSpan period);
-        Task<int> GetOperationsCountForTheLastPeriodAsync(TimeSpan period, Guid userId);
+        Task<long> GetOperationsCountForTheLastPeriodAsync(TimeSpan period);
+        Task<long> GetOperationsCountForTheLastPeriodAsync(TimeSpan period, Guid userId);
         Task<decimal> GetAggregatedAmountForTheLastPeriodAsync(TimeSpan period);
         Task<decimal> GetAggregatedAmountForTheLastPeriodAsync(TimeSpan period, Guid userId);
         Task RegisterOperationAsync(Operation operation);
